if available, use hgattic extension for shelve/unshelve. Else fall back to self-implemented shelve.

+++ src/com/vectrace/MercurialEclipse/operations/UnShelveOperation.java
@@ -24,7 +24,10 @@ import com.vectrace.MercurialEclipse.MercurialEclipsePlugin;
 import com.vectrace.MercurialEclipse.actions.HgOperation;
 import com.vectrace.MercurialEclipse.commands.HgClients;
 import com.vectrace.MercurialEclipse.commands.HgPatchClient;
+import com.vectrace.MercurialEclipse.commands.extensions.HgAtticClient;
 import com.vectrace.MercurialEclipse.exception.HgException;
+import com.vectrace.MercurialEclipse.team.MercurialUtilities;
+import com.vectrace.MercurialEclipse.team.ResourceProperties;
 
 /**
  * @author bastian
@@ -84,32 +87,56 @@ public class UnShelveOperation extends HgOperation {
             // get modified files
             monitor.beginTask(Messages
                     .getString("UnShelveOperation.Unshelving"), 4); //$NON-NLS-1$
-            monitor.subTask(Messages.getString("UnShelveOperation.GettingChanges")); //$NON-NLS-1$
-            File root = HgClients.getHgRoot(project.getLocation().toFile());
-            File shelveDir = new File(root, ".hg" + File.separator //$NON-NLS-1$
-                    + "mercurialeclipse-shelve-backups"); //$NON-NLS-1$
-            
-            if (shelveDir.exists()) {
-                File shelveFile = new File(shelveDir, project.getName().concat(
-                        "-patchfile.patch")); //$NON-NLS-1$
-                if (shelveFile.exists()) {
-                    monitor.worked(1);
-                    monitor.subTask(Messages.getString("UnShelveOperation.applyingChanges")); //$NON-NLS-1$
-                    HgPatchClient.importPatch(project, shelveFile,
-                            new ArrayList<String>(0));
-                    monitor.worked(1);
-                    monitor.subTask(Messages.getString("UnShelveOperation.emptyingShelf")); //$NON-NLS-1$
-                    boolean deleted = shelveFile.delete();
-                    monitor.worked(1);
-                    monitor.subTask(Messages.getString("UnShelveOperation.refreshingProject")); //$NON-NLS-1$
-                    project.refreshLocal(IResource.DEPTH_INFINITE, monitor);
-                    monitor.worked(1);                    
-                    if (!deleted) {
-                        throw new HgException(shelveFile.getName()+" could not be deleted.");
+
+            // check if hg > 1.0x and hgattic is available
+            if (MercurialUtilities.isCommandAvailable("resolve", // $NON-NLS-1$
+                    ResourceProperties.RESOLVE_AVAILABLE, "") // $NON-NLS-1$
+                    && MercurialUtilities.isCommandAvailable("attic-shelve",// $NON-NLS-1$
+                            ResourceProperties.EXT_HGATTIC_AVAILABLE, "")) { // $NON-NLS-1$
+                HgAtticClient.unshelve(project.getLocation().toFile(), false,
+                        project
+                        .getName());
+                monitor.worked(1);
+                project.refreshLocal(IResource.DEPTH_INFINITE, monitor);
+                monitor.worked(1);
+            } else {
+                monitor.subTask(Messages
+                        .getString("UnShelveOperation.GettingChanges")); //$NON-NLS-1$
+                File root = HgClients.getHgRoot(project.getLocation().toFile());
+                File shelveDir = new File(root, ".hg" + File.separator //$NON-NLS-1$
+                        + "mercurialeclipse-shelve-backups"); //$NON-NLS-1$
+
+                if (shelveDir.exists()) {
+                    File shelveFile = new File(shelveDir, project.getName()
+                            .concat("-patchfile.patch")); //$NON-NLS-1$
+                    if (shelveFile.exists()) {
+                        monitor.worked(1);
+                        monitor
+                                .subTask(Messages
+                                        .getString("UnShelveOperation.applyingChanges")); //$NON-NLS-1$
+                        ArrayList<String> opts = new ArrayList<String>();
+                        opts.add("--no-commit");
+                        HgPatchClient.importPatch(project, shelveFile,
+                                opts);
+                        monitor.worked(1);
+                        monitor.subTask(Messages
+                                .getString("UnShelveOperation.emptyingShelf")); //$NON-NLS-1$
+                        boolean deleted = shelveFile.delete();
+                        monitor.worked(1);
+                        monitor
+                                .subTask(Messages
+                                        .getString("UnShelveOperation.refreshingProject")); //$NON-NLS-1$
+                        project.refreshLocal(IResource.DEPTH_INFINITE, monitor);
+                        monitor.worked(1);
+                        if (!deleted) {
+                            throw new HgException(shelveFile.getName()
+                                    + " could not be deleted.");
+                        }
+                    } else {
+                        throw new HgException(
+                                Messages
+                                        .getString("UnShelveOperation.error.ShelfEmpty")); //$NON-NLS-1$
                     }
-                } else {
-                    throw new HgException(
-                            Messages.getString("UnShelveOperation.error.ShelfEmpty")); //$NON-NLS-1$
                 }
             }
         } catch (Exception e) {
